pyproject.toml
[project]
name = "whisper-transcribe"
version = "0.1.0"
description = "A unified API for working with various Whisper speech-to-text implementations"
readme = "README.md"
requires-python = ">=3.10.11"
license = { file = "LICENSE" }
authors = [
    { name = "Rafsun Jany Arman", email = "armanrafsunjany@gmail.com" }
]
keywords = ["whisper", "transcription", "speech-to-text", "audio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Multimedia :: Sound/Audio :: Speech"
]
dependencies = [
    "fastapi>=0.115.11",
    "pydantic-settings>=2.8.1",
    "pyyaml>=6.0.2",
    "uvicorn[standard]>=0.34.0",
]

[project.urls]
"Homepage" = "https://github.com/CrewsBoard/whisper-transcribe"
"Bug Tracker" = "https://github.com/CrewsBoard/whisper-transcribe/issues"
"Documentation" = "https://github.com/CrewsBoard/whisper-transcribe#readme"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest]
testpaths = ["tests"]
